+++ src/test/java/de/gematik/demis/igs/gateway/csv/CsvUploadControllerIT.java
@@ -19,40 +19,52 @@ package de.gematik.demis.igs.gateway.csv;
  * In case of changes by gematik find details in the "Readme" file.
  *
  * See the Licence for the specific language governing permissions and limitations under the Licence.
+ *
+ * *******
+ *
+ * For additional notes and disclaimer from gematik and in case of changes by gematik find details in the "Readme" file.
  * #L%
  */
 
 import static de.gematik.demis.igs.gateway.DefaultExceptionHandler.ERROR_MESSAGE_INVALID_INPUT;
 import static de.gematik.demis.igs.gateway.DefaultExceptionHandler.ERROR_MESSAGE_VALIDATION;
 import static de.gematik.demis.igs.gateway.IgsOverviewCsvTestData.FAKE_DEMIS_NOTIFICATION_ID;
+import static de.gematik.demis.igs.gateway.configuration.MessagesProperties.ERROR_DATE_FORMAT;
+import static de.gematik.demis.igs.gateway.configuration.MessagesProperties.ERROR_DUPLICATE_FILE_NAME;
+import static de.gematik.demis.igs.gateway.configuration.MessagesProperties.ERROR_EMPTY_FILE;
+import static de.gematik.demis.igs.gateway.configuration.MessagesProperties.ERROR_INVALID_FILE_FORMAT;
+import static de.gematik.demis.igs.gateway.configuration.MessagesProperties.ERROR_PREFIX;
+import static de.gematik.demis.igs.gateway.configuration.MessagesProperties.ERROR_PRIME_DIAGNOSTIC_LAB;
+import static de.gematik.demis.igs.gateway.configuration.MessagesProperties.ERROR_REQUIRED;
 import static de.gematik.demis.igs.gateway.csv.OverviewDataExtractor.ERROR_MESSAGE_INCONSISTENT_ROW_LENGTH;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.either;
 import static org.hamcrest.Matchers.everyItem;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import de.gematik.demis.igs.gateway.FutsMockTestTemplate;
 import de.gematik.demis.igs.gateway.IgsOverviewCsvTestData;
+import de.gematik.demis.igs.gateway.configuration.MessageSourceWrapper;
 import de.gematik.demis.igs.gateway.csv.futs.ValueSetMappingService;
 import de.gematik.demis.igs.gateway.csv.model.IgsOverviewCsv;
 import de.gematik.demis.igs.gateway.csv.model.OverviewParsedRowResult;
 import de.gematik.demis.igs.gateway.csv.model.OverviewResponse;
 import de.gematik.demis.igs.gateway.csv.validation.ValidationError.ErrorCode;
-import de.gematik.demis.igs.gateway.csv.validation.ValidationError.ErrorMessage;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Comparator;
 import java.util.List;
 import lombok.SneakyThrows;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
 import org.junit.jupiter.params.provider.NullSource;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -64,14 +76,32 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 @AutoConfigureMockMvc
 class CsvUploadControllerIT extends FutsMockTestTemplate {
 
-  public static final String CONCAT_ERROR_DATE =
-      ErrorMessage.PREFIX.msg() + ErrorMessage.DATE_FORMAT.msg();
-  public static final String CONCAT_DUPLICATE =
-      ErrorMessage.PREFIX.msg() + ErrorMessage.DUPLICATE_FILE_NAME.msg();
-  public static final String CONCAT_REQUIRED =
-      ErrorMessage.PREFIX.msg() + ErrorMessage.REQUIRED.msg();
-  public static final String CONCAT_LAB =
-      ErrorMessage.PREFIX.msg() + ErrorMessage.PRIME_DIAGNOSTIC_LAB.msg();
+  @Autowired ValueSetMappingService valueSetMappingService;
+  @Autowired private MockMvc mockMvc;
+  @Autowired private ObjectMapper objectMapper;
+  @Autowired private MessageSourceWrapper messageSourceWrapper;
+
+  private String CONCAT_ERROR_DATE;
+  private String CONCAT_DUPLICATE;
+  private String CONCAT_REQUIRED;
+  private String CONCAT_LAB;
+
+  @BeforeEach
+  void setup() {
+    CONCAT_ERROR_DATE =
+        messageSourceWrapper.getMessage(ERROR_PREFIX)
+            + messageSourceWrapper.getMessage(ERROR_DATE_FORMAT);
+    CONCAT_DUPLICATE =
+        messageSourceWrapper.getMessage(ERROR_PREFIX)
+            + messageSourceWrapper.getMessage(ERROR_DUPLICATE_FILE_NAME);
+    CONCAT_REQUIRED =
+        messageSourceWrapper.getMessage(ERROR_PREFIX)
+            + messageSourceWrapper.getMessage(ERROR_REQUIRED);
+    CONCAT_LAB =
+        messageSourceWrapper.getMessage(ERROR_PREFIX)
+            + messageSourceWrapper.getMessage(ERROR_PRIME_DIAGNOSTIC_LAB);
+  }
+
   private static final String PATH_TO_CSV =
       "src/test/resources/testdata/igs-batch-fastq-testdata.csv";
   private static final String PATH_TO_MINIMAL_CSV =
@@ -86,17 +116,19 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
       "src/test/resources/testdata/igs-batch-fastq-testdata_empty_rows.csv";
   private static final String PATH_TO_CSV_WITH_WRONG_DATEFORMAT =
       "src/test/resources/testdata/igs-batch-fastq-testdata_wrong_dateformat.csv";
+  private static final String PATH_TO_CSV_WITH_INVALID_FORMAT =
+      "src/test/resources/testdata/igs-batch-fastq-testdata.xlsx";
   private static final String PATH_TO_CSV_WITH_INVALID_PRIME_DIAGNOSTIC_LAB =
       "src/test/resources/testdata/igs-batch-fastq-testdata_invalid_prime_diagnostic_lab.csv";
   private static final String PATH_TO_CSV_WITH_DUPLICATE_FILE_NAMES =
       "src/test/resources/testdata/igs-batch-fastq-duplicate-file-names.csv";
-  @Autowired ValueSetMappingService valueSetMappingService;
-  @Autowired private MockMvc mockMvc;
-  @Autowired private ObjectMapper objectMapper;
+  private static final String PATH_TO_CSV_WITH_LOCAL_DATES =
+      "src/test/resources/testdata/igs-batch-fasta-testdata_local_date_format.csv";
 
-  @Test
-  void shouldUploadCsv() throws Exception {
-    byte[] csvData = Files.readAllBytes(Paths.get(PATH_TO_CSV));
+  @ParameterizedTest
+  @ValueSource(strings = {PATH_TO_CSV, PATH_TO_CSV_WITH_LOCAL_DATES})
+  void shouldUploadCsv(String pathToCsv) throws Exception {
+    byte[] csvData = Files.readAllBytes(Paths.get(pathToCsv));
     valueSetMappingService.refreshCachedValues();
     MvcResult mvcResult =
         mockMvc
@@ -157,7 +189,9 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
         .andExpect(jsonPath("$.title").value(ERROR_MESSAGE_INVALID_INPUT))
         .andExpect(jsonPath("$.errorReport").isArray())
         .andExpect(jsonPath("$.errorReport", hasSize(1)))
-        .andExpect(jsonPath("$.errorReport[0].msg").value(ErrorMessage.EMPTY_FILE.msg()))
+        .andExpect(
+            jsonPath("$.errorReport[0].msg")
+                .value(messageSourceWrapper.getMessage(ERROR_EMPTY_FILE)))
         .andExpect(jsonPath("$.errorReport[0].errorCode").value(ErrorCode.EMPTY_FILE.name()))
         .andExpect(jsonPath("$.errorReport[0].rowNumber").value(0))
         .andExpect(jsonPath("$.errorReport[0].foundValue").isEmpty())
@@ -177,7 +211,9 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
         .andExpect(jsonPath("$.errorReport").isArray())
         .andExpect(jsonPath("$.errorReport", hasSize(1)))
         .andExpect(jsonPath("$.errorReport", hasSize(1)))
-        .andExpect(jsonPath("$.errorReport[0].msg").value(ErrorMessage.EMPTY_FILE.msg()))
+        .andExpect(
+            jsonPath("$.errorReport[0].msg")
+                .value((messageSourceWrapper.getMessage(ERROR_EMPTY_FILE))))
         .andExpect(jsonPath("$.errorReport[0].errorCode").value(ErrorCode.EMPTY_FILE.name()))
         .andExpect(jsonPath("$.errorReport[0].rowNumber").value(0))
         .andExpect(jsonPath("$.errorReport[0].foundValue").isEmpty())
@@ -190,7 +226,6 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
     MvcResult mvcResult =
         mockMvc
             .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
-            .andDo(print())
             .andExpect(status().isOk())
             .andReturn();
 
@@ -217,17 +252,33 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
         .andExpect(jsonPath("$.errorReport", hasSize(1)))
         .andExpect(
             jsonPath("$.errorReport[0].msg")
-                .value(CONCAT_ERROR_DATE.formatted(1, "DATE_OF_RECEIVING", "03.03.2023")))
+                .value(CONCAT_ERROR_DATE.formatted(1, "DATE_OF_RECEIVING", "2023/03/03")))
         .andExpect(jsonPath("$.errorReport[0].errorCode").value(ErrorCode.DATE_FORMAT.name()))
         .andExpect(jsonPath("$.errorReport[0].rowNumber").value(1))
-        .andExpect(jsonPath("$.errorReport[0].foundValue").value("03.03.2023"))
+        .andExpect(jsonPath("$.errorReport[0].foundValue").value("2023/03/03"))
         .andExpect(jsonPath("$.errorReport[0].columnName").value("DATE_OF_RECEIVING"));
+  }
+
+  @Test
+  void shouldReturnBadRequestOnNoSupportedFileFormat() throws Exception {
+    byte[] csvData = Files.readAllBytes(Paths.get(PATH_TO_CSV_WITH_INVALID_FORMAT));
+    mockMvc
+        .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
+        .andExpect(status().isBadRequest())
+        .andExpect(jsonPath("$.title").value(ERROR_MESSAGE_INVALID_INPUT))
+        .andExpect(jsonPath("$.detail").value(ERROR_MESSAGE_VALIDATION))
+        .andExpect(jsonPath("$.errorReport", hasSize(1)))
+        .andExpect(
+            jsonPath("$.errorReport[0].msg")
+                .value(messageSourceWrapper.getMessage(ERROR_INVALID_FILE_FORMAT)))
+        .andExpect(
+            jsonPath("$.errorReport[0].errorCode").value(ErrorCode.INVALID_FILE_FORMAT.name()))
+        .andExpect(jsonPath("$.errorReport[0].rowNumber").value(0));
   }
 
   @Test
   void shouldReturnBadRequestOnCorruptedCsv() throws Exception {
     byte[] csvData = Files.readAllBytes(Paths.get(PATH_TO_CORRUPTED_CSV));
-
     mockMvc
         .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
         .andExpect(status().isBadRequest())
@@ -255,7 +306,6 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
   @Test
   void shouldReturnBadRequestMixedErrorsCsv() throws Exception {
     byte[] csvData = Files.readAllBytes(Paths.get(PATH_TO_MIXED_ERRORS_CSV));
-
     mockMvc
         .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
         .andExpect(status().isBadRequest())
@@ -274,7 +324,6 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
   @Test
   void shouldReturnBadRequestOnCsvWithInvalidPrimeDiagnosticLab() throws Exception {
     byte[] csvData = Files.readAllBytes(Paths.get(PATH_TO_CSV_WITH_INVALID_PRIME_DIAGNOSTIC_LAB));
-
     mockMvc
         .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
         .andExpect(status().isBadRequest())
@@ -294,7 +343,6 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
   @Test
   void shouldReturnBadRequestOnCsvWithDuplicateFilenames() throws Exception {
     byte[] csvData = Files.readAllBytes(Paths.get(PATH_TO_CSV_WITH_DUPLICATE_FILE_NAMES));
-
     mockMvc
         .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
         .andExpect(status().isBadRequest())
@@ -310,15 +358,21 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
 
   @Test
   @SneakyThrows
-  void shouldAddIllegalHeaderToErrorReport() {
+  void shouldRecognizeIllegalHeaderAsInvalidCsv() {
     final String pathToCsv = "src/test/resources/testdata/igs-batch-fasta-illegal-header.csv";
     final byte[] csvData = Files.readAllBytes(Paths.get(pathToCsv));
-
     mockMvc
         .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
         .andExpect(status().isBadRequest())
-        .andExpect(jsonPath("$.errorReport").isArray())
-        .andExpect(jsonPath("$.errorReport", hasSize(1)));
+        .andExpect(jsonPath("$.title").value(ERROR_MESSAGE_INVALID_INPUT))
+        .andExpect(jsonPath("$.detail").value(ERROR_MESSAGE_VALIDATION))
+        .andExpect(jsonPath("$.errorReport", hasSize(1)))
+        .andExpect(
+            jsonPath("$.errorReport[0].msg")
+                .value(messageSourceWrapper.getMessage((ERROR_INVALID_FILE_FORMAT))))
+        .andExpect(
+            jsonPath("$.errorReport[0].errorCode").value(ErrorCode.INVALID_FILE_FORMAT.name()))
+        .andExpect(jsonPath("$.errorReport[0].rowNumber").value(0));
   }
 
   @Test
@@ -327,7 +381,6 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
     final String pathToCsv =
         "src/test/resources/testdata/igs-batch-fasta-illegal-header-and-missing-required-value.csv";
     final byte[] csvData = Files.readAllBytes(Paths.get(pathToCsv));
-
     mockMvc
         .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
         .andExpect(status().isBadRequest())
@@ -344,7 +397,6 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
   void shouldAddLoincMappingToErrorReport() {
     final String pathToCsv = "src/test/resources/testdata/igs-batch-fasta-unknown-loinc.csv";
     final byte[] csvData = Files.readAllBytes(Paths.get(pathToCsv));
-
     mockMvc
         .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
         .andExpect(status().isBadRequest())
@@ -353,5 +405,21 @@ class CsvUploadControllerIT extends FutsMockTestTemplate {
         .andExpect(jsonPath("$.errorReport").isArray())
         .andExpect(jsonPath("$.errorReport", hasSize(2)))
         .andExpect(jsonPath("$.errorReport[?(@.errorCode == 'LOINC_CODE')]").value(hasSize(2)));
+  }
+
+  @Test
+  @SneakyThrows
+  void shouldWorkCorrectlyOnCommaSeperatedCsv() {
+    final String pathToCsv = "src/test/resources/testdata/igs-batch-fasta-commaCsv.csv";
+    final byte[] csvData = Files.readAllBytes(Paths.get(pathToCsv));
+    mockMvc
+        .perform(MockMvcRequestBuilders.multipart("/csv/upload").file("csvFile", csvData))
+        .andExpect(status().isBadRequest())
+        .andExpect(jsonPath("$.detail").value(ERROR_MESSAGE_VALIDATION))
+        .andExpect(jsonPath("$.title").value(ERROR_MESSAGE_INVALID_INPUT))
+        .andExpect(jsonPath("$.errorReport").isArray())
+        .andExpect(jsonPath("$.errorReport", hasSize(76)))
+        .andExpect(jsonPath("$.errorReport[?(@.errorCode == 'REQUIRED_FIELD')]").value(hasSize(25)))
+        .andExpect(jsonPath("$.errorReport[?(@.errorCode == 'HEADER')]").value(hasSize(51)));
   }
 }