+++ ReleaseNotes.md
@@ -2,5 +2,16 @@
 
 # Release Notes IGS-Gateway
 
+## Release 1.2.0
+- implemented centralized message service
+- improved error handling for upload CSV
+
+## Release 1.1.4
+- Updated ospo-resources for adding additional notes and disclaimer
+- setting new ressources in helm chart
+- setting new timeouts and retries in helm chart
+- change base chart to istio hostnames
+- updating dependencies
+
 ## Release 1.1.1
 - First official GitHub-Release